pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel>=0.41"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "har-mlops-pipeline"
version = "2.1.0"
description = "MLOps-Enhanced HAR Pipeline for Anxiety Detection Using Wearable Sensors"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Shalin Vachheta", email = "shalinvachheta2016@hotmail.com"},
]
keywords = [
    "mlops", "human-activity-recognition", "anxiety-detection",
    "wearable-sensors", "deep-learning", "domain-adaptation",
    "uncertainty-quantification", "drift-detection",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    # Core data wrangling
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.10",

    # Deep Learning
    "tensorflow>=2.14",

    # Metrics & evaluation
    "scikit-learn>=1.3",

    # Experiment tracking & MLOps
    "mlflow>=2.11",
    "dvc>=3.50",

    # API & Serving
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.5",
    "python-multipart>=0.0.6",

    # Visualization
    "matplotlib>=3.8",
    "seaborn>=0.13",

    # Utilities
    "tqdm>=4.66",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "pytest-xdist>=3.5",
    "flake8>=7.0",
    "black>=24.1",
    "isort>=5.13",
    "mypy>=1.8",
    "jupyterlab>=4.1",
    "python-dotenv>=1.0",
]
monitoring = [
    "prometheus-client>=0.20",
]
gpu = [
    "tensorflow[and-cuda]>=2.14",
]

[project.scripts]
har-pipeline = "src.cli:main"
update-progress = "scripts.update_progress_dashboard:main"

[project.urls]
Repository = "https://github.com/shalini/MasterArbeit_MLops"
Documentation = "docs/"

# ============================================================================
# Tool Configuration
# ============================================================================

[tool.setuptools.packages.find]
include = ["src*"]

[tool.setuptools.package-data]
src = ["*.yaml", "*.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers -ra"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "gpu: marks tests that require GPU",
    "robustness: marks robustness/noise injection tests",
    "calibration: marks calibration/uncertainty tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/__init__.py", "*/config.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.",
    "pass",
]
